Adding Rules to Ambassador's ClusterRole
We are upgrading EKS to 1.20 & ambassador requires more/different ingress permissions.

Unsure of whether to replace the other ingress rule or add it as I have done in the first commit of this PR.

Testing the upgrade process in DS Stage cluster is where I found this issue. I resolved it by manually adding the following to the ClusterRole yaml.

`- apiGroups:
  - extensions
  - networking.k8s.io
  resources:
  - ingresses
  - ingressclasses
  verbs:
  - get
  - list
  - watch
`